+++ plugins/commentable-html/dev/tests/_validate_helpers.py
@@ -46,6 +46,7 @@ EXPECTED_REQUIRED_IDS = frozenset({
     "btnToggleSidebar", "btnCopyAll", "btnCopyAllTop", "btnClearAll",
     "btnCloseSidebar", "menuComment",
     "btnToolbarMenu", "toolbarMenu",
+    "btnMoreMenu", "sidebarMoreMenu", "btnSearchToggle",
     "btnSaveHtml", "btnSaveHtmlTop", "btnSavePlain", "btnSavePlainTop",
     "btnExportOffline", "btnExportOfflineTop",
     "btnStorage", "btnStorageTop",